# decompression performance improvements (~50%)