--- a/lz4.js
+++ b/lz4.js
@@ -96,9 +96,7 @@ function makeBuffer (size) {
 
 function sliceArray (array, start, end) {
   if (typeof array.buffer !== undefined) {
-    if (Uint8Array.prototype.subarray) {
-      return array.subarray(start, end);
-    }else if (Uint8Array.prototype.slice) {
+    if (Uint8Array.prototype.slice) {
       return array.slice(start, end);
     } else {
       // Uint8Array#slice polyfill.