Two main reasons for perf gains:
1. use [TypedArray.copyWithin ](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray/copyWithin)
2. use [TypedArray.subarray](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray/subarray) which reuses buffer instead of [TypedArray.slice](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray/slice) which makes a copy (and thus more work + GC)

Decompress a 12MB file, representing 100MB of raw data
orig: 230ms
copyWithin15+: 145ms
subarray: 130ms
copyWithin31+: 120ms
lz4cat: 70ms