@jchv 2018-06-02
Ah, I missed this - looks pretty good.

Do you think the `hasCopyWithin` check is sufficient? Do any browsers have partial support for these functions? I am pretty sure the last time I was messing around, I found that browser support for Typed Arrays was oddly inconsistent.

In any case, I'm interested in merging this either way.